+++ tests/cloud_tasks/cli/test_run_argv.py
@@ -99,7 +99,15 @@ def test_run_argv_show_queue_success(tmp_path: Path, capsys: pytest.CaptureFixtu
 def test_run_argv_show_queue_detail_success(
     tmp_path: Path, capsys: pytest.CaptureFixture[str]
 ) -> None:
-    """show_queue --detail with mocked queue and one message."""
+    """show_queue --detail with mocked queue and one message.
+
+    Parameters:
+        tmp_path: Path – temporary directory fixture.
+        capsys: pytest.CaptureFixture[str] – captures stdout/stderr.
+
+    Returns:
+        None.
+    """
     config_path = tmp_path / "config.yaml"
     config_path.write_text("provider: gcp\ngcp:\n  job_id: test-job\n  project_id: test-project\n")
     mock_queue = AsyncMock()
@@ -952,16 +960,13 @@ def test_run_argv_stop_with_purge_queue(tmp_path: Path, capsys: pytest.CaptureFi
     mock_orch.queue_name = "test-job"
     mock_orch.initialize = AsyncMock()
     mock_orch.stop = AsyncMock()
-    mock_task_queue = AsyncMock()
-    mock_task_queue.purge_queue = AsyncMock()
-    mock_orch.task_queue = mock_task_queue
-    mock_orch._task_queue = mock_task_queue  # CLI uses _task_queue for purge
+    mock_orch.purge_queue = AsyncMock()
     with patch("cloud_tasks.cli.InstanceOrchestrator", return_value=mock_orch):
         code = run_argv(
             ["stop", "--config", str(config_path), "--provider", "gcp", "--purge-queue"]
         )
     assert code == 0
-    mock_task_queue.purge_queue.assert_called_once()
+    mock_orch.purge_queue.assert_called_once()
 
 
 def test_run_argv_stop_raises_exits_one(tmp_path: Path) -> None:
